restructured the plugin to align with wc structure and modularized it
