+++ includes/constants.php
@@ -1,0 +1,12 @@
+<?php
+//Define Constants
+define('WC_PAYNOW_VERSION', '1.2.0');
+define('ps_error', 'error');
+define('ps_ok','ok');
+define('ps_created_but_not_paid','created but not paid');
+define('ps_cancelled','cancelled');
+define('ps_failed','failed');
+define('ps_paid','paid');
+define('ps_awaiting_delivery','awaiting delivery');
+define('ps_delivered','delivered');
+define('ps_awaiting_redirect','awaiting redirect');
